+++ src/protean/port/cache.py
@@ -1,7 +1,7 @@
 from abc import ABCMeta, abstractmethod
 from typing import Optional, Union
 
-from protean.core.view import BaseView
+from protean.core.projection import BaseProjection
 from protean.utils.inflection import underscore
 
 
@@ -15,13 +15,13 @@ class BaseCache(metaclass=ABCMeta):
         # Default TTL: 300 seconds
         self.ttl = conn_info.get("TTL", 300)
 
-        # Temporary cache of views
-        self._views = {}
+        # Temporary cache of projections
+        self._projections = {}
 
-    def register_view(self, view_cls):
-        """Registers a view object for data serialization and de-serialization"""
-        view_name = underscore(view_cls.__name__)
-        self._views[view_name] = view_cls
+    def register_projection(self, projection_cls):
+        """Registers a projection object for data serialization and de-serialization"""
+        projection_name = underscore(projection_cls.__name__)
+        self._projections[projection_name] = projection_cls
 
     @abstractmethod
     def ping(self):
@@ -32,21 +32,22 @@ class BaseCache(metaclass=ABCMeta):
         """Get the connection object for the cache"""
 
     @abstractmethod
-    def add(self, view: BaseView, ttl: Optional[Union[int, float]] = None) -> None:
-        """Add view record to cache
+    def add(
+        self, projection: BaseProjection, ttl: Optional[Union[int, float]] = None
+    ) -> None:
+        """Add projection record to cache
 
-        KEY: View ID
-        Value: View Data (derived from `to_dict()`)
+        KEY: Projection ID
+        Value: Projection Data (derived from `to_dict()`)
 
         TTL is in seconds. If not specified explicitly in method call,
         it can be picked up from broker configuration. In the absence of
         configuration, it can be defaulted to an optimum value, say 300 seconds.
 
         Args:
-            view (BaseView): View Instance containing data
+            projection (BaseProjection): Projection Instance containing data
             ttl (int, float, optional): Timeout in seconds. Defaults to None.
         """
-        """Add a key-value to the cache through the view object"""
 
     @abstractmethod
     def get(self, key):
@@ -61,8 +62,8 @@ class BaseCache(metaclass=ABCMeta):
         """Retrieve count of data by key pattern"""
 
     @abstractmethod
-    def remove(self, view):
-        """Remove a cache record by view"""
+    def remove(self, projection):
+        """Remove a cache record by projection object"""
 
     @abstractmethod
     def remove_by_key(self, key):