+++ tests/test_registry.py
@@ -103,7 +103,7 @@ def test_properties_method_returns_a_dictionary_of_all_protean_elements():
         "repositories": DomainObjects.REPOSITORY.value,
         "subscribers": DomainObjects.SUBSCRIBER.value,
         "value_objects": DomainObjects.VALUE_OBJECT.value,
-        "views": DomainObjects.VIEW.value,
+        "projections": DomainObjects.PROJECTION.value,
     }